Renaming the package and config
